agent: fix MCP tools not loaded in headless mode
`mcp::start()` returned the handle before servers finished connecting. The actual init ran on a detached task, so every caller that read tools right after `start()` got an empty `ToolIndex`. Print mode always hit this race.

Inlined the init into `start_with_config()` so it awaits `start_enabled()` before returning. If you have a handle, it is ready. Also moved `shutdown_mcp` into `mcp::shutdown()` so print mode can clean up child processes too (it never did before).
